Move TikTok parser package under search namespace
