+++ tests/services/tiktok/test_tiktok_sigi_state.py
@@ -4,7 +4,7 @@ from typing import Dict, Any
 
 import pytest
 
-from app.services.tiktok.parser.json_parser import _from_sigi_state
+from app.services.tiktok.search.parser.json_parser import _from_sigi_state
 
 
 def _wrap_sigi_state(data: Dict[str, Any]) -> str:
